+++ src/federation/callback.ts
@@ -61,10 +61,24 @@ export type ObjectDispatcher<
 /**
  * A callback that dispatches a collection.
  *
- * @typeParam TContextData The context data to pass to the {@link Context}.
+ * @typeParam TItem The type of items in the collection.
+ * @typeParam TContext The type of the context. {@link Context} or
+ *                     {@link RequestContext}.
+ * @typeParam TContextData The context data to pass to the `TContext`.
+ * @typeParam TFilter The type of the filter, if any.
+ * @param context The context.
+ * @param handle The handle of the collection owner.
+ * @param cursor The cursor to start the collection from, or `null` to dispatch
+ *               the entire collection without pagination.
+ * @param filter The filter to apply to the collection, if any.
  */
-export type CollectionDispatcher<TItem, TContextData, TFilter> = (
-  context: RequestContext<TContextData>,
+export type CollectionDispatcher<
+  TItem,
+  TContext extends Context<TContextData>,
+  TContextData,
+  TFilter,
+> = (
+  context: TContext,
   handle: string,
   cursor: string | null,
   filter?: TFilter,
@@ -84,10 +98,20 @@ export type CollectionCounter<TContextData, TFilter> = (
 /**
  * A callback that returns a cursor for a collection.
  *
+ * @typeParam TContext The type of the context. {@link Context} or
+ *                     {@link RequestContext}.
  * @typeParam TContextData The context data to pass to the {@link Context}.
+ * @typeParam TFilter The type of the filter, if any.
+ * @param context The context.
+ * @param handle The handle of the collection owner.
+ * @param filter The filter to apply to the collection, if any.
  */
-export type CollectionCursor<TContextData, TFilter> = (
-  context: RequestContext<TContextData>,
+export type CollectionCursor<
+  TContext extends Context<TContextData>,
+  TContextData,
+  TFilter,
+> = (
+  context: TContext,
   handle: string,
   filter?: TFilter,
 ) => string | null | Promise<string | null>;