+++ src/federation/context.ts
@@ -202,6 +202,22 @@ export interface Context<TContextData> {
     activity: Activity,
     options?: SendActivityOptions,
   ): Promise<void>;
+
+  /**
+   * Sends an activity to the outboxes of the sender's followers.
+   * @param sender The sender's handle.
+   * @param recipients In this case, it must be `"followers"`.
+   * @param activity The activity to send.
+   * @param options Options for sending the activity.
+   * @throws {Error} If no followers collection is registered.
+   * @since 0.14.0
+   */
+  sendActivity(
+    sender: { handle: string },
+    recipients: "followers",
+    activity: Activity,
+    options?: SendActivityOptions,
+  ): Promise<void>;
 }
 
 /**
@@ -271,36 +287,6 @@ export interface RequestContext<TContextData> extends Context<TContextData> {
    * @since 0.7.0
    */
   getSignedKeyOwner(): Promise<Actor | null>;
-
-  /**
-   * Sends an activity to recipients' inboxes.
-   * @param sender The sender's handle or the sender's key pair(s).
-   * @param recipients The recipients of the activity.
-   * @param activity The activity to send.
-   * @param options Options for sending the activity.
-   */
-  sendActivity(
-    sender: SenderKeyPair | SenderKeyPair[] | { handle: string },
-    recipients: Recipient | Recipient[],
-    activity: Activity,
-    options?: SendActivityOptions,
-  ): Promise<void>;
-
-  /**
-   * Sends an activity to the outboxes of the sender's followers.
-   * @param sender The sender's handle.
-   * @param recipients In this case, it must be `"followers"`.
-   * @param activity The activity to send.
-   * @param options Options for sending the activity.
-   * @throws {Error} If no followers collection is registered.
-   * @since 0.8.0
-   */
-  sendActivity(
-    sender: { handle: string },
-    recipients: "followers",
-    activity: Activity,
-    options?: SendActivityOptions,
-  ): Promise<void>;
 }
 
 /**